requirements.txt
Flask==0.9
Flask-WTF==0.8.2
Jinja2==2.6
WTForms==1.0.2
Werkzeug==0.8.3
argparse==1.2.1
distribute==0.6.24
factory-boy==1.2.0
flask-mongoengine==0.6
mongoengine==0.7.8
nose==1.2.1
pymongo==2.4.1
wsgiref==0.1.2
